# openapi.yaml
openapi: 3.0.3
info:
  title: cloudup REST API
  description: >
    Local REST API for cloudup, a parallel-upload manager for cloud storage
    (Google Drive, S3 and S3-compatible, WebDAV/Yandex.Disk). Runs as a
    local server (see cmd/server); any frontend, including the reference
    Vue app, talks to it over this API. Progress and queue state are
    polled, not pushed - see the /tasks endpoints.
  version: "1.0"
servers:
  - url: http://127.0.0.1:3000
security:
  - bearerAuth: []
paths:
...
  /api/v1/provider-types/{type}/schema:
    get:
      summary: Connection form fields for a provider type
      description: >
        Used to render an "add connection" form generically, without
        hardcoding a form per provider type - mirrors
        internal/provider.ConfigSchema.
      parameters:
        - $ref: "#/components/parameters/providerType"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FieldSpec"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/connections:
    get:
      summary: List connections
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      summary: Create a connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/connections/{id}:
    get:
      summary: Get a connection
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a connection's display name and fields
      description: >
        secrets, if present, only overwrites the given keys - existing
        secret fields not mentioned are left untouched. providerType and
        createdAt are immutable.
      parameters:
        - $ref: "#/components/parameters/connectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a connection
      description: >
        Also best-effort deletes every secret field (keychain entries) this
        connection owns. Refuses to delete a connection with a pending or
        in-flight upload (409) - cancel those first via
        POST /api/v1/connections/{id}/cancel-all - or with an enabled watch
        rule still pointing at it (409) - disable or delete that rule
        first via PUT/DELETE /api/v1/watches/{id}.
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The connection has pending or in-flight uploads.
  /api/v1/connections/{id}/test:
    post:
      summary: Test a connection
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "200":
          description: >
            Always 200 - inspect body.ok for the actual result, since a
            failed test is an expected outcome, not a server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  error:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/connections/{id}/uploads:
    post:
      summary: Upload a file through this connection
      description: >
        Spools the uploaded file to local disk and enqueues it on
        internal/queue.Manager (FIFO per connection, parallel across
        connections, automatic retry). Poll /api/v1/tasks/{taskId} for
        progress.
      parameters:
        - $ref: "#/components/parameters/connectionId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                remotePath:
                  type: string
                  description: Defaults to the uploaded file's name if omitted.
              required:
                - file
      responses:
        "202":
          description: Accepted - queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tasks:
    get:
      summary: List known upload tasks (poll for progress)
      parameters:
        - name: connectionId
          in: query
          description: Filter to a single connection's tasks.
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /api/v1/tasks/{id}:
    get:
      summary: Get a single task's current state
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tasks/{id}/cancel:
    post:
      summary: Cancel a task, pending or in flight
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: Cancel requested
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/connections/{id}/pause:
    post:
      summary: Pause a connection's upload queue
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "204":
          description: Paused
  /api/v1/connections/{id}/resume:
    post:
      summary: Resume a connection's upload queue
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "204":
          description: Resumed
  /api/v1/connections/{id}/cancel-all:
    post:
      summary: Cancel every pending and in-flight task for a connection
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "204":
          description: Cancel requested
  /api/v1/connections/{id}/oauth/authorize:
    post:
      summary: Start the interactive OAuth2 consent flow for this connection
      description: >
        Only for connections whose provider type has requiresOAuth set (see
        GET /api/v1/provider-types); the app-wide client credentials for that
        type must be configured first (PUT
        /api/v1/provider-types/{type}/oauth-credentials), which must
        already have this server's own GET /api/v1/oauth/callback
        registered as their redirect URI (see that endpoint's description).
        Returns the consent URL the user must open; poll the GET variant of
        this endpoint for completion - the actual token exchange happens
        when the provider redirects the browser to GET
        /api/v1/oauth/callback, not synchronously with this call.
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "202":
          description: Flow started
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
        "400":
          description: >
            The connection's provider type needs no interactive
            authorization, or its client credentials are not configured yet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      summary: Poll the status of an in-progress (or just-finished) authorize flow
      parameters:
        - $ref: "#/components/parameters/connectionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
                  done:
                    type: boolean
                  error:
                    type: string
        "404":
          description: No authorize flow has been started for this connection
  /api/v1/provider-types/{type}/oauth-credentials:
    get:
      summary: Whether this provider type's app-wide OAuth client is configured
      description: >
        Reports presence only - the stored values are never returned.
      parameters:
        - $ref: "#/components/parameters/providerType"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
        "400":
          description: This provider type needs no interactive authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Configure this provider type's app-wide OAuth Client ID/Secret
      description: >
        One registered OAuth client per provider type (a Google Cloud "Web
        application" client, a Dropbox App) covers every connection of that
        type - set once, shared by all of them (see
        internal/provider.OAuthFlow.AppCredentialsID). Values are stored in
        the OS keychain, never in config.json, and are never readable back
        through this API. Whichever redirect URI you register with the
        provider must be this server's own GET /api/v1/oauth/callback,
        reachable however clients actually reach this API (e.g.
        http://127.0.0.1:3000/api/v1/oauth/callback for the default local
        run, or your real public URL behind a reverse proxy).
      parameters:
        - $ref: "#/components/parameters/providerType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId, clientSecret]
              properties:
                clientId:
                  type: string
                clientSecret:
                  type: string
      responses:
        "204":
          description: Stored
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/oauth/callback:
    get:
      summary: OAuth2 redirect target - not called directly by API clients
      security: []
      description: >
        This is the exact URL to register as every provider's redirect URI
        (see PUT /api/v1/provider-types/{type}/oauth-credentials) - the
        provider's own consent page redirects the user's browser here after
        they approve or deny access, so unlike every other /api/v1/* route
        it is unauthenticated: the browser arriving from Google/Dropbox/
        Yandex/Microsoft has no bearer token to attach. state ties the
        request back to whichever connection's POST .../oauth/authorize
        started it (rejecting anything it doesn't recognize, including a
        replayed callback - each state is single-use); code is exchanged
        for a refresh token, stored under that connection. Renders a small
        HTML page either way, since a real browser lands here - poll
        GET /api/v1/connections/{id}/oauth/authorize for the machine-
        readable outcome instead of parsing this response.
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
        - name: error
          in: query
          description: Present instead of code when the user denied access
          schema:
            type: string
      responses:
        "200":
          description: >
            An HTML page reporting success or failure - always 200 once
            state is recognized, since the actual outcome belongs in the
            authorize-status poll, not this response's status code.
          content:
            text/html:
              schema:
                type: string
        "400":
          description: Unknown, expired, or already-used state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/history:
    get:
      summary: List upload history entries
      description: >
        Paginated, newest upload first. Defaults to 50 entries per page
        (limit's default and maximum are the server's
        DefaultHistoryPageSize/MaxHistoryPageSize) - a long-running
        install's log can grow large over time, so this endpoint never
        returns it whole in one response.
      parameters:
        - name: connectionId
          in: query
          schema:
            type: string
        - name: status
          in: query
          description: One of success / error / cancelled.
          schema:
            type: string
        - name: limit
          in: query
          description: >
            Max entries to return. Defaults to 50; values above 200 are
            silently clamped to 200 - see the response's `limit` field for
            what was actually used.
          schema:
            type: integer
        - name: offset
          in: query
          description: How many matching entries to skip before this page. Defaults to 0.
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
        "400":
          description: limit or offset is not a valid integer.
  /api/v1/history/{id}:
    get:
      summary: Get a single history entry
      parameters:
        - $ref: "#/components/parameters/historyId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a history entry
      description: Only removes the local journal row - never touches the remote object.
      parameters:
        - $ref: "#/components/parameters/historyId"
      responses:
        "204":
          description: Deleted
  /api/v1/history/{id}/verify:
    post:
      summary: Re-check that an uploaded object still exists and matches its checksum
      description: >
        Sets the returned HistoryEntry's LastCheckStatus: "missing" if the
        object is gone, "mismatch" if it exists but the checksum no longer
        matches, "ok" if it matches, "unverifiable" if the provider
        supports neither an existence nor a checksum check, or "error" if
        reaching the provider itself failed.
      parameters:
        - $ref: "#/components/parameters/historyId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Required on every /api/v1/* request. The token is generated randomly
        at each server startup unless CLOUDUP_API_TOKEN is set, and is never
        written to disk. Where to obtain it depends on how the server runs:
        when it serves the bundled frontend (the default) the token is
        injected into index.html as window.__CLOUDUP_TOKEN__, so a browser
        opening the page is already authorized and nothing is logged; in
        API-only mode (-static "") it is written to the server log instead,
        since there is no page to inject it into.
  parameters:
...
    providerType:
      name: type
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    FieldSpec:
      type: object
      properties:
        Key:
          type: string
        Label:
          type: string
        Type:
          type: string
          enum: [text, password, select]
        Required:
          type: boolean
        Options:
          type: array
          items:
            type: string
...